JAMES-3138 Move RecomputeQuotaService to a mailbox-tools-quota-recompute module
This allows removing data-api from mailbox-store dependencies.
